Set a conversion rule

This function constructs the conversion between two landuse types or a commodity and a landuse type.

Usage

setConversion(
  from = NULL,
  to = NULL,
  label = NULL,
  thresh = NULL,
  auto = NULL,
  mask = NULL
)

Arguments

from

character(1)
a landuse type for which a relationship shall be created.

to

character(1)
a landuse type to which the relationship shall be created.

label

character(1)
description of the relationship.

thresh

integerish(1)
the number of time-steps the landuse type in from must remain the same before it can change into the landuse type in to.

auto

integerish(1)
the number of time-steps after which the landuse type in from will automatically change into to.

mask

character(1)
the name of a gridded object for which conversion to the specified class is allowed.

Details

From the resulting land-use change sequence, a land-use conversion matrix is derived.

See also