Set a location specific preference

Set a location specific preference

Usage

setPreference(landuse = NULL, layer = NULL, weight = NULL, table = NULL)

Arguments

landuse

character(1)
the landuse type for which to specify location specific preference.

layer

character(1)
the gridded object in which the location specific preference values are noted. If this is left empty, it will be a layer that includes only the value 0 for all pixels.

weight

numeric(1)
the weight (beween -1 for down-weighing and 1 for up-weighing the preferences) assigned to this factor.

table

data.frame(1)
table of information on location specific preference. Must contain the columns landuse, layer, and weight.

subset

character(1)
an expression that successfully evaluates to the subset of a raster, for example to subset a layer on the fly between years.

Details

layer = "self" is reserved to derive dynamic locspec maps. In that case the age map of each timestep is multiplied by weight. Depending on the sign of weight, either all age-value of this class (negative) weight) or of all other classes (positive weight) are selected for multiplication (all others are set to 0). De facto, this leads to decrease of preference where the class already occurs with a negative sign, and increase of preference where the class does not yet exist with a positive sign. This can be interessting, when a class becomes less likely, the longer it is present, e.g., in crop rotation schemes, where both, ressource depletion or pathogen pressure but also increasing times of commodities that repair the soil, make rotation more likely.

See also