Start a new project

This function registers initial metadata of the current project into the global options.

Usage

startProject(
  root = NULL,
  name = NULL,
  start = NULL,
  end = NULL,
  landuse = NULL,
  demands = NULL,
  resolution = NULL,
  extent = NULL
)

Arguments

root

character(1)
path to the root directory that contains or shall contain the working directory of this project.

name

character(1)
the name for this project.

start

integerish(1)
the first year of the project.

end

integerish(1)
the last year of the project.

landuse

character(.)
the allowed/valid landuse systems for this project.

demands

character(.)
the allowed/valid demand types for this project.

resolution

numeric(2)
vector of x and y resolution.

extent

extent(1)
maximum extent of the region, raster::extent object.

Value

Enters the provided information as 'lclu' into the options.